handle rest api requests with edit context as admin requests
With Gutenberg dynamic blocks are requested through the rest api.
If those blocks contain images, currently these images will be replaced with the placeholder and not display correctly.

This commit checks if we are in the rest api and if the current context is edit. If so, it will not replace the images.